start/apps/images/urls.py
from django.conf.urls import url, include
from rest_framework.routers import DefaultRouter
from .views import ImageViewSet

router = DefaultRouter()
router.register(r'images', ImageViewSet)

urlpatterns = [
    url('', include(router.urls))
]
