requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.5
django-coverage-plugin==1.5.0

Sphinx==1.6.6
django-extensions==1.9.9
Werkzeug==0.14.1
django-test-plus==1.0.22
factory-boy==2.10.0

django-debug-toolbar==1.9.1

# improved REPL
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.9.0

django-rest-swagger==2.1.2
